continuing updating branch to solve conflicts with dev branch
